requirements.txt
python-dotenv~=0.20.0
Flask~=2.2.1
APScheduler~=3.9.1
requests~=2.28.1